refactor: launch argument
